gRPC/maps_service/requirements.txt
fastapi
uvicorn[standard]
grpcio
grpcio-tools
protobuf
jinja2
